refactor(setup_page): move setup steps to separated widgets
